Job running scheduler

+++ src/main/java/it/aredegalli/printer/dto/job/request/JobStartRequestDto.java
@@ -9,6 +9,6 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class JobStartRequestDto {
 
-    private Integer startOffsetLine;
+    private Long startOffsetLine;
 
 }
