+++ src/main/java/it/aredegalli/printer/dto/job/request/JobUpdateRequestDto.java
@@ -9,6 +9,6 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class JobUpdateRequestDto {
 
-    private Integer progress;
+    private Long progress;
 
 }
